+++ navground_sim_py/navground/sim/list_plugins.py
@@ -1,0 +1,14 @@
+import argparse
+
+from navground import sim
+from navground.core.list_plugins import list_plugins, parser
+
+
+def _main(arg: argparse.Namespace) -> None:
+    sim.load_plugins()
+    list_plugins(arg, sim.get_loaded_plugins())
+
+
+def main() -> None:
+    arg = parser().parse_args()
+    _main(arg)